Make HDMI audio work more reliably, especially in current aosp build
In code ported from ADT-1, there was an optimization not to update
HDMI routing when a new connection was enabled/disabled if the
current outputs were not changed from the new one.  Problem is
that audioHAL by default considered HDMI not connected, but
the audio policy says it was.

This bug wasn't noticed in regular full build because at some point
in boot, another output device is added (a2dp in my log) which caused
the updateRoute to be called, but for some reason that other output
isn't triggering in aosp_fugu (I think because bluetooth is off by
default there).

Anyway, fix the bug by getting rid of the optimization and always
update the route, so HDMI state is in sync between audio policy manager
and audio HAL.

Change-Id: I2b1ffc8160dd3c7867362c487abd953c20afd5f7
Signed-off-by: Mike J. Chen <mjchen@google.com>