wifi: adjust 'tcp_limit_output_bytes' to increase TCP upload speed
In the last version of the WIFI driver, the number of transmit
buffers has been increased from 256 to 1000, hence set
'tcp_limit_output_bytes' to reflect this.
At the same time, set the 'tcp_rmem' and 'tcp_wmem' sizes to the same
amount, in order to not stop/start transmit queues.
Tested with Chariot:
- TCP UL throughput -- Initial : around 320Mbps
- TCP UL throughput -- With the patch: around 420Mbps

Bug: 17386901
BZ: 220921
Change-Id: I266e6debb13f34abda01f4fabd45883323e1658e
Signed-off-by: Jonathan DE CESCO <jonathan.de.cesco@intel.com>
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>