don't use logwrapper for dhcp services
When init is computing the SELinux domain for a service,
it examines the file label of the executable. This causes
problems when processes are wrapped using logwrapper; SELinux
is unable to determine the appropriate security domain for a
process, and ends up running it in the wrong domain.

Remove the logwrapper call to avoid this problem.

Fixes the following dmesg warning:

  <11>[  162.377220] init: Warning!  Service dhcpcd_wlan0 needs a SELinux domain defined; please fix!

Change-Id: If80e765f777db8861540cad5f2fdbc481b839b65