Temporarily disable GPU Renderscript IntrinsicResize
Temporarily disable GPU Renderscript IntrinsicResize by (in
powervr.ini) specifying keyword DisableIntrinsics with the bitmap value
4096 (which is 1 << 12), where RS_SCRIPT_INTRINSIC_ID_RESIZE = 12.

(Note: Previously DisableIntrinsics=1 was specified, which is a no-op.)

This action is taken due to failures from CTS test class
android.renderscript.cts.IntrinsicResize and due to the schedule being
close enough to the upcoming MR1 milestone to cause the more proper fix
to be deemed too risky at this time.  It is contemplated that after MR1
this commit will be reverted and replaced by the solution described
below.

A patch is available now to make the computation accurate to within
1 LSB.  In conjunction with a proposed patch to the IntrinsizeResize
tests to allow the 1 LSB deviation, the test issue would be resolved.

Bug: 18706159
BZ: 230436
Change-Id: I3109cc7b34c8709a9773ec383de4ee8b99f843c7
Signed-off-by: Dale Stimson <dale.b.stimson@intel.com>