Retry NTP indefinitely until successful
On a cold boot, the RTC will not be set.  If there are network
issues (lan is up but wan is down), the default value of 3 could
cause the device to not get NTP in a timely fashion.
(by default, after 3 failures, the device would wait 24h to try
again)

Bug: 19046107

Change-Id: I4015f359853d8d1e7a47e8e1c20bc086a1d0cb13
Signed-off-by: Patrick Tjin <pattjin@google.com>