Set support for A2DP to false
Pairing of A2DP devices is added in ag/636110. The change allows pairing of
audio devices only if A2DP support is true. fugu is a fixed volume
device and it does not support A2DP. Therefore, we are setting it to false.

The change was tested on the mr1-release branch.

Change-Id: I14f22cfe7a33389ff4c2380aaa8f029ba089868a