Fix for the jittery issue because of mismatch in buffer size
The data written by application should match the buffer size
configured in the driver or it should be multiple of the buffer
size configured. If not during the getHardwareTimestamp query,
some time you will get residual bytes as avail buffer. This
will be for a fraction of millisecond, which need to be ignored.

BUG=19604395

Change-Id: Ie503dbed30c3662d50e2ad9d52dd88370f1c7c59
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>