Remove COMMON_GLOBAL_CFLAGS
This isn't currently used in any code that I can find. Even if it was,
the individual modules where it is used should set it in their
individual cflags.

We're removing the ability to change COMMON_GLOBAL_* in the product/device
configurations.

COMMON_GLOBAL_CFLAGS affects both host and device cflags, so switching
between devices that change this variable would cause all the host tools
to be rebuilt unnecessarily.

Bug: 26165350
Change-Id: I4a703ee106eaa3cc87cf3d30a7538edd10bd1161