gpu: Updates to support DDK 1.6.
Removed vp9_interpredict.binary. This is a shader binary which has a
layout specific to the GPU DDK revision. It should be provided with
the DDK.

Removed the pvrsrvctl binary. The GPU is started automatically at
the first open() now and the firmware is loaded by the kernel.

The sepolicy needed minor tweaks for cameraservice. I removed the
permissive rules for the obsolete pvrsrvctl binary to close that
security hole.

Removed most of the powervr.ini settings. Most of these were already
the default, or have been removed from the new DDK.

Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>
Bug: 26592113
Bug: 27643307
Change-Id: Ie20fad85c0373d3ae0510afff65827e8a5ea3ab6