dexpreopt: dont preopt prebuilts and select apps
Preoptimizing apps pushes compile-time from first-boot to system image
generation and thus improves out-of-the-box experience, at the cost of
system image space.

As part of disabling squashfs on the system parition, we don't preopt
prebuilt apps that are likely to be updated from the Play Store.

The apps listed to not be preoptimized were found by running:
grep odex $ANDROID_PRODUCT_OUT/installed-files.txt
which lists the size of odex files sorted from largest in size to
smallest.

Apps were selected based on their size and subjective importantance to
remove preopt-ing

Bug: 29360840
Change-Id: I74d6cde05d9efcf836992c7ea1b20cfa7ac20102